% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utility.R
\name{fromBinary}
\alias{fromBinary}
\title{Convert a vector of 1s and 0s to an integer}
\usage{
fromBinary(v)
}
\arguments{
\item{v}{an \code{integer} vector}
}
\value{
an integer value
}
\description{
Turns a vector of 1s and 0s representing class membership into an integer
}

