% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utility.R
\name{toBinary}
\alias{toBinary}
\title{Convert integer to vector of 1s and 0s}
\usage{
toBinary(y)
}
\arguments{
\item{y}{an integer}
}
\value{
a \code{numeric} vector of 1s and 0s
}
\description{
Turns integer split representation into vector of class membership
}

