% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/translate.R
\name{translate}
\alias{translate}
\title{Generate translated code from various R tree models}
\usage{
translate(obj, lang = "sas", pfx, n)
}
\arguments{
\item{obj}{a tree-based model object (e.g. gbm, randomForest)}

\item{lang}{language to translate object into (e.g. 'sas', 'ecl')}

\item{pfx}{string to prepend to each tree prediction 'pfx_TREE_0000N'}

\item{n}{number of trees to print. Prints \code{1:n} tree predictions}
}
\value{
returns a character vector containing all of the translated text
}
\description{
This function generates translated code of a specified language and returns
it in a character vector.
}
\details{
The generated code is written to standard output.  See example for how
to write to a file.
}

